chatbot-knowledge-base/requirements.txt
fastapi==0.103.2
langchain==0.0.312
pydantic==2.4.2
pydantic_settings==2.0.3
PyJWT==2.8.0
python-dotenv==1.0.0
qdrant_client==1.6.0
uvicorn==0.23.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
openai==0.28.1